+++ src/app/pages/login/login.page.ts
@@ -1,14 +1,11 @@
-import { Component, OnInit }         from '@angular/core';
-import { Router }                    from '@angular/router';
+import { Component, OnInit }           from '@angular/core';
+import { NgForm }                      from '@angular/forms';
+import { Router }                      from '@angular/router';
 
-import { AlertController, Platform } from '@ionic/angular';
+import { LoadingController, Platform } from '@ionic/angular';
 
-import { AuthService }               from '@app-services/auth/auth.service';
-import { UserService }               from '@app-services/user/user.service';
-
-
-const EMAIL_REGEXP: RegExp = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-const PASS_REGEXP: RegExp = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{6,16}$/;
+import { AuthService, EMAIL_REGEXP, PASS_REGEXP } from '@app-services/auth/auth.service';
+import { UserService }                 from '@app-services/user/user.service';
 
 
 @Component({
@@ -18,11 +15,12 @@ const PASS_REGEXP: RegExp = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{6,16}$/;
 })
 export class LoginPage implements OnInit {
 
-  public email: string;
-  public password: string;
+  public emailPattern: RegExp = EMAIL_REGEXP;
+  public passwordPattern: RegExp = PASS_REGEXP;
+  public passwordVisible: boolean = false;
 
-  constructor(private platform: Platform, 
-              private alertCtrl: AlertController,
+  constructor(private loadingCtrl: LoadingController,
+              private platform: Platform, 
               private auth: AuthService,
               private users: UserService,
               private router: Router) { }
@@ -30,82 +28,68 @@ export class LoginPage implements OnInit {
   public ngOnInit(): void {
   }
 
-  public async googleLogin() {
+  public async googleLogin(): Promise<void> {
     try {
-      if (this.platform.is('cordova')) {
-        const credential = await this.auth.googleLogin();
-        if (credential.user) {
-          console.log('user: ', credential.user);
-          this.router.navigate(['/']);
-        }
-        else console.log('user error'); 
+      const credential = (this.platform.is('cordova')) ? 
+        await this.auth.googleLogin() :
+        await this.auth.browserSocialLogin('google');
+
+      if (credential.user) {
+        console.log('user: ', credential.user);
+        this.users.updateUser(credential.user);
+        this.router.navigateByUrl('/tabs/(map:map)');
       }
-      else {
-        const credential = await this.auth.browserSocialLogin('google');
-        if (credential.user) {
-          console.log('user: ', credential.user);
-          this.users.updateUser(credential.user);
-          this.router.navigateByUrl('/tabs/(map:map)');
-        }
-        else console.log('user error');
-      }
+      else console.log('user error'); 
     }
     catch(e) { console.log('googleBrowserLogin() error: ', e.message) }
   }
 
-  public async doLogin(): Promise<void> {
+  public async facebookLogin(): Promise<void> {
     try {
-      if (this.checkEmail(this.email)) {
-        if (this.checkPassword(this.password)) {
-          const credential = await this.auth.emailLogin(this.email, this.password);
-          if (credential.user) {
-            console.log('user: ', credential.user);
-            this.users.updateUser(credential.user);
-            this.router.navigateByUrl('/tabs/(map:map)');
-          }
-          else console.error('user error');
-        }
-        else {
-          const errorMessage: string = 'Please re/enter a valid password';
-          await this.showErrorAlert(errorMessage);
-        }
+      const credential = (this.platform.is('cordova')) ?
+        await this.auth.facebookLogin() :
+        await this.auth.browserSocialLogin('facebook');
+        
+      if (credential.user) {
+        console.log('user: ', credential.user);
+        this.users.updateUser(credential.user);
+        this.router.navigateByUrl('/tabs/(map:map)');
       }
-      else {
-        const errorMessage: string = 'Please enter a valid email';
-        await this.showErrorAlert(errorMessage);
+      else console.log('user error'); 
+    }
+    catch(e) { console.error('facebookLogin() error: ', e) }
+  }
+
+  public async doLogin(form: NgForm): Promise<void> {
+    let loader: HTMLIonLoadingElement;
+
+    try {
+      const email: string = form.value.email;
+      const password: string = form.value.password;
+      loader = await this.loadingCtrl.create({ message: 'Authenticating...', spinner: 'circles' });
+
+      await loader.present();
+
+      const credential = await this.auth.emailLogin(email, password);
+      if (credential.user) {
+        console.log('user: ', credential.user);
+        this.users.updateUser(credential.user);
+        loader.dismiss();
+        this.router.navigateByUrl('/tabs/(map:map)');
       }
     }
-    catch(e) { console.error('doLogin() error: ', e) }
-  }
-
-  private checkEmail(email: string): boolean {
-    const e = email.trim();
-    console.log('trimmed: ', e);
-
-    if (e.length < 8) return false;
-    else if (!EMAIL_REGEXP.test(e)) return false;
-    else return true;
-  }
-
-  private checkPassword(password: string): boolean {
-    const p: string = password.trim();
-    console.log('trimmed: ', p);
-    return PASS_REGEXP.test(p);
-  }
-
-  private async showErrorAlert(msg: string): Promise<void> {
-    try {
-      const errorAlert = await this.alertCtrl.create({
-        header: 'Register Error',
-        message: msg,
-        buttons: ['OK']
-      });
-      await errorAlert.present();
+    catch(e) { 
+      if(loader != undefined) loader.dismiss();
+      //console.log('doLogin() error: ', e);
     }
-    catch(e) { console.error('showErrorAlert() error: ', e) }
-    
+  }
+
+  public async doPasswordReset() {
+    try { await this.auth.showResetPasswordAlert() }
+    catch(e) { console.error('doPasswordReset() error: ', e) }
   }
 
   get isAndroid(): boolean { return this.platform.is('android') }
 
+  public togglePasswordVisiblity(): void { this.passwordVisible = !this.passwordVisible }
 }